ENH: GetArrayFromImage and GetImageFromArray perform deep copy
This implementation of GetArrayFromImage() and GetImageFromArray() functions
perform a deep copy of the array so that the memory in the source object
and in the target objects are different. This avoids potential side-effects
of using a memory view. The implementation follows the implementation
of GetArrayFromImage() in SimpleITK [1].

[1] https://github.com/SimpleITK/SimpleITK/blob/next/Wrapping/Python/\
Python.i#L745-L755