ENH: Remove Copy-On-Write which was not consistent across all conversions
A copy-on-write was performed when converting ITK objects (Image, VNL vectors,
VNL Matrices) to NumPy array, and the copy-on-write operation was only
performed when the NumPy array was modified. This means that if the ITK object
was modified, no copy-on-write was performed, and therefore the NumPy array
would contain the modifications of the ITK object.
To remove these inconsistencies, copy-on-write is removed, and user must
either use the deep-copy functions or the views.